+++ CHANGELOG.md
@@ -9,6 +9,7 @@
     * Added roaring import support for `RowKeyColumnID`, `RowIDColumnKey` and `RowKeyColumnKey` type data. Pass `OptImportRoaring(true)` to `client.ImportField` to activate that.
     * Field option getters are attached to `Field` type instead of `*Field` type.
     * Deprecated `field.Options`. Use `field.Opts` instead.
+    * Deprecated `Range` call. Use `RowRange` instead.
 
 * **v1.2.0** (2018-12-21)
     * **Compatible with Pilosa 1.2**